reference-node/pyproject.toml
[project]
name = "openposter-reference-node"
version = "0.1.0"
description = "OpenPoster reference node (FastAPI)"
requires-python = ">=3.12"
dependencies = [
  "fastapi>=0.115",
  "uvicorn[standard]>=0.30",
  "pydantic>=2.7",
  "sqlalchemy>=2.0",
  "aiosqlite>=0.20",
  "pynacl>=1.5",
  "cryptography>=42.0",
  "rfc8785>=0.1.4",
]

[project.optional-dependencies]
dev = [
  "pytest>=8.0",
  "httpx>=0.27",
]

[tool.pytest.ini_options]
asyncio_mode = "auto"

[tool.uv]
